Reorganize lock icons
