+++ packages/ui/src/app/pages/Profile/components/LockItem.tsx
@@ -8,7 +8,7 @@ import { ButtonGhost } from '@/common/components/buttons'
 import { DropDownButton, DropDownToggle } from '@/common/components/buttons/DropDownToggle'
 import { Arrow } from '@/common/components/icons'
 import { BoxIcon } from '@/common/components/icons/BoxIcon'
-import { VoteIcon } from '@/common/components/icons/VoteIcon'
+import { VoteIcon } from '@/common/components/icons/locks'
 import { Label, TextMedium, TextSmall, TokenValue } from '@/common/components/typography'
 import { BorderRad, Colors } from '@/common/constants'
 import { useToggle } from '@/common/hooks/useToggle'
