+++ packages/ui/src/common/components/icons/locks/EnvelopeIcon.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Icon } from './Icon'
+import { Icon } from '../Icon'
 
 interface EnvelopeIconProps {
   className?: any
