+++ packages/ui/src/common/components/icons/locks/VoteIcon.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Icon } from './Icon'
+import { Icon } from '../Icon'
 
 interface VoteIconProps {
   className?: any
