+++ packages/ui/src/common/components/icons/locks/index.ts
@@ -1,0 +1,4 @@
+export * from './EnvelopeIcon'
+export * from './LabelIcon'
+export * from './LockIcon'
+export * from './VoteIcon'
